requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
drf-yasg
django-filter==22.1
djangorestframework-simplejwt==5.2.2
